move out of ui.js
